+++ algorithms/0035-search-insert-position.py
@@ -1,6 +1,12 @@
 class Solution:
-    def searchInsert(self, nums: List[int], target: int) -> int:
-        for i in range(len(nums)):
-            if nums[i] >= target:
-                return i
-        return len(nums)
+    def searchInsert(self, nums, target):
+        l, r = 0, len(nums)
+        while l < r:
+            mid = (l + r) // 2
+            if nums[mid] == target:
+                return mid
+            elif nums[mid] < target:
+                l = mid + 1
+            else:
+                r = mid
+        return r